// src/pedidos/pedido-estado.enum.ts
export enum EstadoPedido {
    SOLICITADO = 'solicitado',
    CON_VARIACIONES = 'con_variaciones',
    APROBADO = 'aprobado',
    EN_VALIDACION = 'en_validacion',
    EN_PREPARACION = 'en_preparacion',
    EN_CAMINO = 'en_camino',
    ENTREGADO = 'entregado',
    FINALIZADO = 'finalizado',
    CANCELADO = 'cancelado',
    RECHAZADO = 'rechazado',

  }